Safari-609.1.11.1/LayoutTests/platform/ios/imported/w3c/web-platform-tests/custom-elements/reactions/HTMLInputElement-expected.txt

CONSOLE MESSAGE: line 14: TypeError: The HTMLInputElement.capture getter can only be used on instances of HTMLInputElement

FAIL Custom Elements: CEReactions on HTMLInputElement interface TypeError: The HTMLInputElement.capture getter can only be used on instances of HTMLInputElement
